Elixir: Exercise "German Sysadmin"; Refactored; Documented
The code was modified to no longer use special characters as magic constants.
In addition, the cases were modified to try to make them more readable.
Finally, the sanitization flow was changed to no longer use reduce (foldr in this case).
The use of filters and flat_map was preferred, since this flow is more intuitive.